Botpharos/requirements.txt
web3
requests
python-dotenv
random-user-agent
rich
